+++ src/hip_attn/v1_2/attention_extend.py
@@ -27,12 +27,12 @@ from hip_attn.v1_2.attention_metadata import (
     ScanStage,
     safe_stride,
 )
-
-# from hip_attn.v1_2.scan_stage import (
-#     chunk_controllable_sampling_mask_cuda,
-#     load_keys_with_rope,
-# )
-from hip_attn.v1_2.uvm_gpu_cache import load_tokens
+from hip_attn.v1_2.compute_v_cos import compute_v_cos
+from hip_attn.v1_2.eval_stage import calculate_chunk_score
+from hip_attn.v1_2.scan_stage import (
+    chunk_controllable_sampling_mask_cuda,
+    load_keys_with_rope,
+)
 
 _NUM_STREAMING_MULTIPROCESSOR = None
 
@@ -115,1861 +115,6 @@ def render_plot_ks(indices, ks, debug, DEBUG_HEAD, BLOCK_SIZE_Q):
 
 DEBUG = os.getenv("HIP_DEBUG", "0") == "1"
 DEBUG_RENDER = os.getenv("HIP_DEBUG_RENDER", "1") == "1"
-
-
-@triton.jit
-def load_keys_with_rope(
-    K,
-    stride_k_bsz,
-    stride_k_tsrc,
-    stride_k_head_kv,
-    stride_k_hid,
-    COS,
-    stride_cos_t,
-    stride_cos_hid,
-    SIN,
-    stride_sin_t,
-    stride_sin_hid,
-    # paged attention args template
-    USING_PAGES,
-    PAGE_SIZE,
-    K_CACHE,
-    stride_k_cache_page,
-    stride_k_cache_offset,
-    stride_k_cache_kv_head,
-    stride_k_cache_hid,
-    BLOCK_TABLE,
-    stride_block_table_bsz,
-    stride_block_table_page,
-    CACHE_SEQ_LENS,
-    stride_cache_seq_lens_b,
-    USING_OFFLOAD_CACHE: tl.constexpr,
-    OFFLOAD_CACHE_KV_PACKED: tl.constexpr,
-    GPU_BANK_COUNT,
-    OFFLOAD_CACHE_UVM_METADATA,
-    stride_offload_cache_uvm_metadata_token,
-    stride_offload_cache_uvm_metadata_k,
-    OFFLOAD_CACHE_GPU_GLOBAL_METADATA,
-    stride_offload_cache_gpu_global_metadata_k,
-    stride_offload_cache_gpu_global_metadata_pad,
-    OFFLOAD_CACHE_GPU_BANK,
-    stride_offload_cache_gpu_bank_token,
-    stride_offload_cache_gpu_bank_hid,
-    OFFLOAD_CACHE_GPU_METADATA,
-    stride_offload_cache_gpu_metadata_token,
-    stride_offload_cache_gpu_metadata_k,
-    OFFLOAD_CACHE_GPU_TABLE,
-    stride_offload_cache_gpu_table_head_kv,
-    stride_offload_cache_gpu_table_token,
-    strdie_offload_cache_gpu_table_k,
-    ACCESS_COUNTER,
-    stride_access_counter_bsz,
-    stride_access_counter_head_kv,
-    stride_access_counter_tsrc,
-    CACHE_MISS_COUNTER,
-    stride_cache_miss_counter_bsz,
-    stride_cache_miss_counter_head_kv,
-    stride_cache_miss_counter_tsrc,
-    queries,
-    idx_bsz,
-    idx_tsrc,
-    idx_head_kv,
-    idx_hid,
-    idx_chunk,
-    mask_tsrc_active,
-    mask_tdst,
-    mask_hid,
-    real_pos_tdst_min,
-    model_context_length,
-    num_sinks,
-    USING_EXTEND,
-    EXTEND_BACKEND,
-    NEED_APPLY_ROPE,
-    BLOCK_CHUNK,
-    BLOCK_HID: tl.constexpr,
-    HID_DIM,
-    IS_RIGHT,
-    HEAD_KV,
-    UPDATE_CACHE,
-    rope_range_begin: tl.constexpr,
-    rope_range_end: tl.constexpr,
-):
-    idx_rope_range = idx_hid - rope_range_begin
-    rope_mask = (rope_range_begin <= idx_hid) & (idx_hid < rope_range_end)
-    ROPE_DIM = rope_range_end - rope_range_begin
-
-    keys_left = load_tokens(
-        K,
-        stride_k_bsz,
-        stride_k_tsrc,
-        stride_k_head_kv,
-        stride_k_hid,
-        USING_PAGES,
-        PAGE_SIZE,
-        K_CACHE,
-        stride_k_cache_page,
-        stride_k_cache_offset,
-        stride_k_cache_kv_head,
-        stride_k_cache_hid,
-        BLOCK_TABLE,
-        stride_block_table_bsz,
-        stride_block_table_page,
-        CACHE_SEQ_LENS,
-        stride_cache_seq_lens_b,
-        USING_OFFLOAD_CACHE,
-        OFFLOAD_CACHE_KV_PACKED,
-        GPU_BANK_COUNT,
-        False,
-        OFFLOAD_CACHE_UVM_METADATA,
-        stride_offload_cache_uvm_metadata_token,
-        stride_offload_cache_uvm_metadata_k,
-        OFFLOAD_CACHE_GPU_GLOBAL_METADATA,
-        stride_offload_cache_gpu_global_metadata_k,
-        stride_offload_cache_gpu_global_metadata_pad,
-        OFFLOAD_CACHE_GPU_BANK,
-        stride_offload_cache_gpu_bank_token,
-        stride_offload_cache_gpu_bank_hid,
-        OFFLOAD_CACHE_GPU_METADATA,
-        stride_offload_cache_gpu_metadata_token,
-        stride_offload_cache_gpu_metadata_k,
-        OFFLOAD_CACHE_GPU_TABLE,
-        stride_offload_cache_gpu_table_head_kv,
-        stride_offload_cache_gpu_table_token,
-        strdie_offload_cache_gpu_table_k,
-        ACCESS_COUNTER,
-        stride_access_counter_bsz,
-        stride_access_counter_head_kv,
-        stride_access_counter_tsrc,
-        CACHE_MISS_COUNTER,
-        stride_cache_miss_counter_bsz,
-        stride_cache_miss_counter_head_kv,
-        stride_cache_miss_counter_tsrc,
-        idx_bsz,
-        idx_tsrc[None, :],
-        idx_head_kv,
-        idx_hid[:, None],
-        mask_tsrc_active[None, :],  # & mask_hid[:, None],
-        # mask_tsrc_active[None, :] & mask_hid[:, None],
-        HEAD_KV,
-        BLOCK_CHUNK,
-        BLOCK_HID,
-        HID_DIM,
-        UPDATE_CACHE=UPDATE_CACHE,
-    ).to(queries.dtype)
-
-    if USING_EXTEND:
-        real_pos_tdst_max = tl.sum(mask_tdst.to(tl.int32)) + real_pos_tdst_min
-        tsrc_extend = tl.maximum(0, real_pos_tdst_max - model_context_length)
-        if NEED_APPLY_ROPE or (tsrc_extend >= 0):
-            old_tsrc = idx_tsrc
-
-            if EXTEND_BACKEND == "dynamic_extend":
-                window = model_context_length // 4
-
-                new_tsrc = tl.where(
-                    (idx_tsrc >= (real_pos_tdst_max - window))
-                    | (real_pos_tdst_max <= model_context_length),
-                    idx_tsrc,
-                    # idx_tsrc * 0 + real_pos_tdst_max,
-                    (
-                        (idx_tsrc.to(tl.float32) - (real_pos_tdst_min - window))
-                        * (
-                            (model_context_length - window)
-                            / (real_pos_tdst_min - window)
-                        ).to(tl.float32)
-                    ).to(tl.int32)
-                    + (real_pos_tdst_min - window),
-                )
-                # new_tsrc = idx_tsrc * 0 + real_pos_tdst_max
-                new_tsrc = tl.maximum(
-                    real_pos_tdst_max - model_context_length, new_tsrc
-                )
-            elif EXTEND_BACKEND == "self_extend":
-                window = 8192
-                group_size = 16
-
-                new_tsrc = tl.where(
-                    idx_tsrc >= (real_pos_tdst_max - window),
-                    idx_tsrc,
-                    tl.where(
-                        real_pos_tdst_max <= model_context_length,
-                        idx_tsrc,
-                        (idx_tsrc - real_pos_tdst_min) // group_size
-                        + real_pos_tdst_min,
-                    ),
-                )
-                new_tsrc = tl.maximum(0, new_tsrc)
-            elif EXTEND_BACKEND == "relative":
-                new_tsrc = idx_chunk * 0
-                if IS_RIGHT:
-                    new_tsrc += 1
-            elif EXTEND_BACKEND == "infllm":
-                new_tsrc = idx_chunk * 0
-            elif EXTEND_BACKEND == "streaming":
-                # streaming
-                new_tsrc = idx_chunk
-            else:
-                raise Exception()
-
-            if not NEED_APPLY_ROPE:
-                tl.static_assert(False)
-                keys_left = keys_left.trans(1, 0)
-                keys_left = adjust_rope(
-                    keys_left,
-                    old_tsrc,
-                    new_tsrc,
-                    mask_tsrc_active,
-                    idx_hid,
-                    COS,
-                    stride_cos_t,
-                    stride_cos_hid,
-                    SIN,
-                    stride_sin_t,
-                    stride_sin_hid,
-                    BLOCK_CHUNK,
-                    BLOCK_HID,
-                    HID_DIM,
-                    NEED_APPLY_ROPE,
-                    rope_range_begin,
-                    rope_range_end,
-                ).to(keys_left.dtype)
-                keys_left = tl.trans(keys_left, 1, 0)
-                keys_left = (keys_left * mask_tsrc_active[None, :]).to(keys_left.dtype)
-            else:
-                rope_rot_idx = tl.where(
-                    rope_mask,
-                    (idx_hid - rope_range_begin + ROPE_DIM // 2) % ROPE_DIM
-                    + rope_range_begin,
-                    idx_hid,
-                )
-                keys_left_rot = load_tokens(
-                    K,
-                    stride_k_bsz,
-                    stride_k_tsrc,
-                    stride_k_head_kv,
-                    stride_k_hid,
-                    USING_PAGES,
-                    PAGE_SIZE,
-                    K_CACHE,
-                    stride_k_cache_page,
-                    stride_k_cache_offset,
-                    stride_k_cache_kv_head,
-                    stride_k_cache_hid,
-                    BLOCK_TABLE,
-                    stride_block_table_bsz,
-                    stride_block_table_page,
-                    CACHE_SEQ_LENS,
-                    stride_cache_seq_lens_b,
-                    USING_OFFLOAD_CACHE,
-                    OFFLOAD_CACHE_KV_PACKED,
-                    GPU_BANK_COUNT,
-                    False,
-                    OFFLOAD_CACHE_UVM_METADATA,
-                    stride_offload_cache_uvm_metadata_token,
-                    stride_offload_cache_uvm_metadata_k,
-                    OFFLOAD_CACHE_GPU_GLOBAL_METADATA,
-                    stride_offload_cache_gpu_global_metadata_k,
-                    stride_offload_cache_gpu_global_metadata_pad,
-                    OFFLOAD_CACHE_GPU_BANK,
-                    stride_offload_cache_gpu_bank_token,
-                    stride_offload_cache_gpu_bank_hid,
-                    OFFLOAD_CACHE_GPU_METADATA,
-                    stride_offload_cache_gpu_metadata_token,
-                    stride_offload_cache_gpu_metadata_k,
-                    OFFLOAD_CACHE_GPU_TABLE,
-                    stride_offload_cache_gpu_table_head_kv,
-                    stride_offload_cache_gpu_table_token,
-                    strdie_offload_cache_gpu_table_k,
-                    ACCESS_COUNTER,
-                    stride_access_counter_bsz,
-                    stride_access_counter_head_kv,
-                    stride_access_counter_tsrc,
-                    CACHE_MISS_COUNTER,
-                    stride_cache_miss_counter_bsz,
-                    stride_cache_miss_counter_head_kv,
-                    stride_cache_miss_counter_tsrc,
-                    idx_bsz,
-                    idx_tsrc[None, :],
-                    idx_head_kv,
-                    rope_rot_idx[:, None],
-                    mask_tsrc_active[None, :],
-                    HEAD_KV,
-                    BLOCK_CHUNK,
-                    BLOCK_HID,
-                    HID_DIM,
-                    # NOTE: in previous load, the fetch should be succesfully done.
-                    UPDATE_CACHE=UPDATE_CACHE,
-                ).to(queries.dtype)
-
-                # TODO: multiply -right
-                # keys_left_rot = tl.where(
-                #     (idx_hid + BLOCK_HID // 2)[:, None] < BLOCK_HID,
-                #     -keys_left_rot,
-                #     keys_left_rot
-                # )
-
-                keys_left_rot *= (
-                    (idx_rope_range + ROPE_DIM // 2 < ROPE_DIM) * (-2) + 1
-                )[:, None].to(keys_left_rot.dtype)
-
-                cos_new = tl.load(
-                    COS
-                    + new_tsrc[None, :].to(tl.int64) * stride_cos_t
-                    + (idx_rope_range % (ROPE_DIM // 2))[:, None] * stride_cos_hid,
-                    mask=mask_tsrc_active[None, :] & rope_mask[:, None],
-                    other=0.0,
-                ).to(keys_left.dtype)
-                sin_new = tl.load(
-                    SIN
-                    + new_tsrc[None, :].to(tl.int64) * stride_sin_t
-                    + (idx_rope_range % (ROPE_DIM // 2))[:, None] * stride_sin_hid,
-                    mask=mask_tsrc_active[None, :] & rope_mask[:, None],
-                    other=0.0,
-                ).to(keys_left.dtype)
-
-                keys_left = tl.where(
-                    rope_mask[:, None],
-                    keys_left * cos_new + keys_left_rot * sin_new,
-                    keys_left,
-                )
-
-    return keys_left
-
-
-@triton.jit
-def chunk_controllable_sampling_mask_cuda(
-    Q,
-    stride_q_bsz,
-    stride_q_tdst,
-    stride_q_head,
-    stride_q_hid,
-    K,
-    stride_k_bsz,
-    stride_k_tsrc,
-    stride_k_head_kv,
-    stride_k_hid,
-    POS,
-    stride_pos_bsz,
-    stride_pos_tdst,
-    # paged attention args template
-    USING_PAGES: tl.constexpr,
-    PAGE_SIZE: tl.constexpr,
-    K_CACHE,
-    stride_k_cache_page,
-    stride_k_cache_offset,
-    stride_k_cache_kv_head,
-    stride_k_cache_hid,
-    V_CACHE,
-    stride_v_cache_page,
-    stride_v_cache_offset,
-    stride_v_cache_kv_head,
-    stride_v_cache_hid,
-    BLOCK_TABLE,
-    stride_block_table_bsz,
-    stride_block_table_page,
-    CACHE_SEQ_LENS,
-    stride_cache_seq_lens_b,
-    USING_OFFLOAD_CACHE: tl.constexpr,
-    OFFLOAD_CACHE_KV_PACKED: tl.constexpr,
-    GPU_BANK_COUNT,
-    OFFLOAD_CACHE_UVM_METADATA,
-    stride_offload_cache_uvm_metadata_token,
-    stride_offload_cache_uvm_metadata_k,
-    OFFLOAD_CACHE_GPU_GLOBAL_METADATA,
-    stride_offload_cache_gpu_global_metadata_k,
-    stride_offload_cache_gpu_global_metadata_pad,
-    OFFLOAD_CACHE_GPU_BANK,
-    stride_offload_cache_gpu_bank_token,
-    stride_offload_cache_gpu_bank_hid,
-    OFFLOAD_CACHE_GPU_METADATA,
-    stride_offload_cache_gpu_metadata_token,
-    stride_offload_cache_gpu_metadata_k,
-    OFFLOAD_CACHE_GPU_TABLE,
-    stride_offload_cache_gpu_table_head_kv,
-    stride_offload_cache_gpu_table_token,
-    strdie_offload_cache_gpu_table_k,
-    INDICES_LEFT,
-    stride_indices_left_bsz,
-    stride_indices_left_bdst,
-    stride_indices_left_head,
-    stride_indices_left_chunk,
-    INDICES_RIGHT,
-    stride_indices_right_bsz,
-    stride_indices_right_bdst,
-    stride_indices_right_head,
-    stride_indices_right_chunk,
-    OUT_SCORES,
-    stride_out_scores_bsz,
-    stride_out_scores_bdst,
-    stride_out_scores_head,
-    stride_out_scores_chunk,
-    COS,
-    stride_cos_t,
-    stride_cos_hid,
-    SIN,
-    stride_sin_t,
-    stride_sin_hid,
-    rope_range_begin: tl.constexpr,
-    rope_range_end: tl.constexpr,
-    MASK_ACCESS_COUNTER,
-    stride_mask_access_counter_bsz,
-    stride_mask_access_counter_head_kv,
-    stride_mask_access_counter_tsrc,
-    MASK_CACHE_MISS_COUNTER,
-    stride_mask_cache_miss_counter_bsz,
-    stride_mask_cache_miss_counter_head_kv,
-    stride_mask_cache_miss_counter_tsrc,
-    CHUNK_COUNT: int,
-    MAX_TSRC: int,
-    TDST: int,
-    HEAD: int,
-    sliding_window_size: int,
-    num_sinks: int,
-    model_context_length: int,
-    group_jobs: int,
-    total_jobs: int,
-    HID_DIM: int,
-    BLOCK_HID: tl.constexpr,
-    BLOCK_SIZE_Q: tl.constexpr = 32,
-    STRIDE_Q: tl.constexpr = 1,
-    BLOCK_CHUNK: tl.constexpr = 32,
-    HEAD_GROUP: tl.constexpr = 4,
-    REDUCE: tl.constexpr = "max",
-    USING_EXTEND: tl.constexpr = False,
-    EXTEND_BACKEND: tl.constexpr = "relative",
-    NEED_APPLY_ROPE: tl.constexpr = False,
-    TERMINATE_SIZE: tl.constexpr = 1,
-    SCAN_STRIDE: tl.constexpr = 1,
-    UPDATE_CACHE: tl.constexpr = True,
-    ORACLE_MAXIMUM: tl.constexpr = False,
-):
-    BDST = tl.cdiv(TDST, BLOCK_SIZE_Q)
-    BDST_SCAN = tl.cdiv(BDST, SCAN_STRIDE)
-    BCHUNK = tl.cdiv(CHUNK_COUNT, BLOCK_CHUNK)
-
-    pid_group = tl.program_id(0).to(tl.int64)
-
-    for i in range(group_jobs):
-        pid = pid_group * group_jobs + i
-        if pid < total_jobs:
-            idx_bdst_scan = pid % BDST_SCAN
-            pid = pid // BDST_SCAN
-            idx_bchunk = pid % BCHUNK
-            pid = pid // BCHUNK
-            idx_head = pid % HEAD
-            pid = pid // HEAD
-            idx_bsz = pid
-
-            # idx_tdst = idx_bdst * BLOCK_SIZE_Q + tl.arange(0, BLOCK_SIZE_Q // STRIDE_Q) * STRIDE_Q
-            # mask_tdst = idx_tdst < TDST
-            if BLOCK_SIZE_Q // STRIDE_Q < 16:
-                idx_tdst = (
-                    (BDST - 1)
-                    - (BDST_SCAN - 1) * SCAN_STRIDE
-                    + idx_bdst_scan * SCAN_STRIDE
-                ) * BLOCK_SIZE_Q + tl.arange(0, 16) * STRIDE_Q
-                mask_tdst = (
-                    (idx_tdst < TDST)
-                    & (idx_tdst >= 0)
-                    & (tl.arange(0, 16) < (BLOCK_SIZE_Q // STRIDE_Q))
-                )
-            else:
-                idx_tdst = (
-                    (BDST - 1)
-                    - (BDST_SCAN - 1) * SCAN_STRIDE
-                    + idx_bdst_scan * SCAN_STRIDE
-                ) * BLOCK_SIZE_Q + tl.arange(0, BLOCK_SIZE_Q // STRIDE_Q) * STRIDE_Q
-                mask_tdst = (idx_tdst < TDST) & (idx_tdst >= 0)
-            idx_hid = tl.arange(0, BLOCK_HID)
-            mask_hid = idx_hid < HID_DIM  # (tl.arange(0, BLOCK_HID) % 4) == 0
-
-            idx_rope_range = idx_hid - rope_range_begin
-            rope_mask = (rope_range_begin <= idx_hid) & (idx_hid < rope_range_end)
-            ROPE_DIM = rope_range_end - rope_range_begin
-
-            pos_tdst = tl.load(
-                POS + idx_bsz * stride_pos_bsz + idx_tdst * stride_pos_tdst,
-                mask=mask_tdst,
-                other=0,
-            )
-
-            # real_pos_tdst_min = idx_bdst * BLOCK_SIZE_Q + TSRC - TDST
-            real_pos_tdst_min = tl.min(tl.where(mask_tdst, pos_tdst, 999999999))
-            real_pos_tdst_min = tl.where(
-                tl.sum(mask_tdst.to(tl.int32)) > 0, real_pos_tdst_min, -1
-            )
-
-            if real_pos_tdst_min >= 0:
-                pos_tdst_min = (real_pos_tdst_min - sliding_window_size).to(tl.int32)
-                pos_tdst_min = tl.maximum(pos_tdst_min, 0)
-
-                idx_chunk = idx_bchunk * BLOCK_CHUNK + tl.arange(0, BLOCK_CHUNK)
-                mask_chunk = idx_chunk < CHUNK_COUNT
-
-                idx_tsrc_left = tl.load(
-                    INDICES_LEFT
-                    + idx_bsz * stride_indices_left_bsz
-                    + idx_bdst_scan * stride_indices_left_bdst
-                    + idx_head * stride_indices_left_head
-                    + idx_chunk * stride_indices_left_chunk,
-                    mask=mask_chunk,
-                    other=MAX_TSRC,
-                ).to(tl.int32)
-
-                idx_tsrc_right = tl.load(
-                    INDICES_RIGHT
-                    + idx_bsz * stride_indices_right_bsz
-                    + idx_bdst_scan * stride_indices_right_bdst
-                    + idx_head * stride_indices_right_head
-                    + idx_chunk * stride_indices_right_chunk,
-                    mask=mask_chunk,
-                    other=MAX_TSRC,
-                ).to(tl.int32)
-
-                if (real_pos_tdst_min + BLOCK_SIZE_Q * SCAN_STRIDE) >= tl.min(
-                    idx_tsrc_left
-                ):
-                    # mask_chunk = mask_chunk & (idx_tsrc_left < (real_pos_tdst_min - sliding_window_size + BLOCK_SIZE_Q))
-
-                    # max_chunk_size = tl.ceil(MAX_TSRC / CHUNK_COUNT).to(tl.float32)
-                    max_chunk_size = tl.max(idx_tsrc_right - idx_tsrc_left).to(
-                        tl.float32
-                    )
-
-                    scores = tl.zeros((BLOCK_CHUNK,), dtype=tl.float32) - 32000.0
-
-                    queries_sum = tl.zeros(
-                        (BLOCK_SIZE_Q // STRIDE_Q, BLOCK_HID), dtype=tl.float32
-                    )
-                    queries_counter = tl.zeros(
-                        (BLOCK_SIZE_Q // STRIDE_Q,), dtype=tl.int32
-                    )
-                    tl.static_assert(BLOCK_SIZE_Q // STRIDE_Q > 0)
-
-                    for i_offset in tl.static_range(STRIDE_Q):
-                        idx_tdst_iter = idx_tdst + i_offset
-                        mask_tdst_iter = mask_tdst & (idx_tdst_iter < TDST)
-                        queries_iter = tl.load(
-                            Q
-                            + idx_bsz * stride_q_bsz
-                            + idx_tdst_iter[:, None] * stride_q_tdst
-                            + idx_head * stride_q_head
-                            + idx_hid[None, :] * stride_q_hid,
-                            mask=mask_tdst_iter[:, None] & mask_hid[None, :],
-                            other=0.0,
-                        )
-                        if queries_iter.dtype == tl.float8e5:
-                            queries_iter = queries_iter.to(tl.float16)
-
-                        if USING_EXTEND:
-                            if NEED_APPLY_ROPE or (
-                                real_pos_tdst_min >= model_context_length
-                            ):
-                                old_tdst = pos_tdst
-                                if EXTEND_BACKEND == "dynamic_extend":
-                                    new_tdst = pos_tdst
-                                elif EXTEND_BACKEND == "self_extend":
-                                    new_tdst = pos_tdst
-                                elif EXTEND_BACKEND == "relative":
-                                    new_tdst = pos_tdst * 0 + 1 + sliding_window_size
-                                elif EXTEND_BACKEND == "infllm":
-                                    new_tdst = pos_tdst * 0 + sliding_window_size
-                                elif EXTEND_BACKEND == "streaming":
-                                    # streaming
-                                    new_tdst = tl.minimum(
-                                        pos_tdst, CHUNK_COUNT + sliding_window_size
-                                    )
-                                else:
-                                    raise Exception()
-
-                                if NEED_APPLY_ROPE:
-                                    rope_rot_idx = (
-                                        idx_hid - rope_range_begin + ROPE_DIM // 2
-                                    ) % ROPE_DIM + rope_range_begin
-                                    queries_rot = tl.load(
-                                        Q
-                                        + idx_bsz * stride_q_bsz
-                                        + idx_tdst_iter[:, None] * stride_q_tdst
-                                        + idx_head * stride_q_head
-                                        + rope_rot_idx[None, :] * stride_q_hid,
-                                        mask=mask_tdst_iter[:, None]
-                                        & rope_mask[None, :]
-                                        & mask_hid[None, :],
-                                        other=0.0,
-                                        # cache_modifier='.cg',
-                                        # eviction_policy='evict_last',
-                                        # volatile=True,
-                                    )
-                                    if queries_rot.dtype == tl.float8e5:
-                                        queries_rot = queries_rot.to(tl.float16)
-
-                                    cos_new = tl.load(
-                                        COS
-                                        + new_tdst[:, None].to(tl.int64) * stride_cos_t
-                                        + (idx_rope_range % (ROPE_DIM // 2))[None, :]
-                                        * stride_cos_hid,
-                                        mask=mask_tdst_iter[:, None]
-                                        & rope_mask[None, :]
-                                        & mask_hid[None, :],
-                                        other=0.0,
-                                    ).to(queries_iter.dtype)
-                                    sin_new = tl.load(
-                                        SIN
-                                        + new_tdst[:, None].to(tl.int64) * stride_sin_t
-                                        + (idx_rope_range % (ROPE_DIM // 2))[None, :]
-                                        * stride_sin_hid,
-                                        mask=mask_tdst_iter[:, None]
-                                        & rope_mask[None, :]
-                                        & mask_hid[None, :],
-                                        other=0.0,
-                                    ).to(queries_iter.dtype)
-
-                                    # queries_rot = tl.where(
-                                    #     (idx_hid + BLOCK_HID // 2)[None, :] < BLOCK_HID,
-                                    #     -queries_rot,
-                                    #     queries_rot
-                                    # )
-                                    queries_rot *= (
-                                        (idx_rope_range + ROPE_DIM // 2 < ROPE_DIM)
-                                        * (-2)
-                                        + 1
-                                    )[None, :].to(queries_rot.dtype)
-
-                                    queries_iter = tl.where(
-                                        rope_mask[None, :] & mask_hid[None, :],
-                                        (
-                                            queries_iter * cos_new
-                                            + queries_rot * sin_new
-                                        ).to(queries_iter.dtype),
-                                        queries_iter,
-                                    )
-                                else:
-                                    queries_iter = adjust_rope(
-                                        queries_iter,
-                                        old_tdst,
-                                        new_tdst,
-                                        mask_tdst_iter,
-                                        idx_hid,
-                                        COS,
-                                        stride_cos_t,
-                                        stride_cos_hid,
-                                        SIN,
-                                        stride_sin_t,
-                                        stride_sin_hid,
-                                        BLOCK_SIZE_Q // STRIDE_Q,
-                                        BLOCK_HID,
-                                        HID_DIM,
-                                        NEED_APPLY_ROPE,
-                                        rope_range_begin,
-                                        rope_range_end,
-                                    ).to(queries_iter.dtype)
-                                    queries_iter = (
-                                        queries_iter * mask_tdst_iter[:, None]
-                                    ).to(queries_iter.dtype)
-
-                        queries_sum += queries_iter
-                        queries_counter += mask_tdst_iter.to(tl.int32)
-
-                    queries = (
-                        queries_sum / (queries_counter[:, None] + 1e-12)
-                    ) * mask_tdst[:, None]
-                    if Q.dtype.element_ty != tl.float8e5:
-                        queries = queries.to(Q.dtype.element_ty)
-                    else:
-                        queries = queries.to(tl.float16)
-
-                    while max_chunk_size >= TERMINATE_SIZE:
-                        max_chunk_size /= 2.0
-                        mask_tsrc_active = (
-                            mask_chunk
-                            & (idx_tsrc_left < idx_tsrc_right)
-                            & (idx_tsrc_left <= pos_tdst_min)
-                            & (idx_tsrc_left >= 0)
-                        )
-                        # mask_tsrc_active = mask_tsrc_active & (idx_tsrc_left < (real_pos_tdst_min - sliding_window_size + BLOCK_SIZE_Q))
-                        idx_tsrc_center = (idx_tsrc_left + idx_tsrc_right) // 2
-
-                        if ORACLE_MAXIMUM:
-                            scores_left = tl.full(
-                                (idx_tdst.shape[0], BLOCK_CHUNK),
-                                dtype=tl.float32,
-                                value=float("-inf"),
-                            )
-                            for tt in range(0, max_chunk_size.to(tl.int32)):
-                                idx_tsrc = idx_tsrc_left + tt
-                                keys_left = load_keys_with_rope(
-                                    K,
-                                    stride_k_bsz,
-                                    stride_k_tsrc,
-                                    stride_k_head_kv,
-                                    stride_k_hid,
-                                    COS,
-                                    stride_cos_t,
-                                    stride_cos_hid,
-                                    SIN,
-                                    stride_sin_t,
-                                    stride_sin_hid,
-                                    # paged attention args template
-                                    USING_PAGES,
-                                    PAGE_SIZE,
-                                    K_CACHE,
-                                    stride_k_cache_page,
-                                    stride_k_cache_offset,
-                                    stride_k_cache_kv_head,
-                                    stride_k_cache_hid,
-                                    BLOCK_TABLE,
-                                    stride_block_table_bsz,
-                                    stride_block_table_page,
-                                    CACHE_SEQ_LENS,
-                                    stride_cache_seq_lens_b,
-                                    USING_OFFLOAD_CACHE,
-                                    OFFLOAD_CACHE_KV_PACKED,
-                                    GPU_BANK_COUNT,
-                                    OFFLOAD_CACHE_UVM_METADATA,
-                                    stride_offload_cache_uvm_metadata_token,
-                                    stride_offload_cache_uvm_metadata_k,
-                                    OFFLOAD_CACHE_GPU_GLOBAL_METADATA,
-                                    stride_offload_cache_gpu_global_metadata_k,
-                                    stride_offload_cache_gpu_global_metadata_pad,
-                                    OFFLOAD_CACHE_GPU_BANK,
-                                    stride_offload_cache_gpu_bank_token,
-                                    stride_offload_cache_gpu_bank_hid,
-                                    OFFLOAD_CACHE_GPU_METADATA,
-                                    stride_offload_cache_gpu_metadata_token,
-                                    stride_offload_cache_gpu_metadata_k,
-                                    OFFLOAD_CACHE_GPU_TABLE,
-                                    stride_offload_cache_gpu_table_head_kv,
-                                    stride_offload_cache_gpu_table_token,
-                                    strdie_offload_cache_gpu_table_k,
-                                    MASK_ACCESS_COUNTER,
-                                    stride_mask_access_counter_bsz,
-                                    stride_mask_access_counter_head_kv,
-                                    stride_mask_access_counter_tsrc,
-                                    MASK_CACHE_MISS_COUNTER,
-                                    stride_mask_cache_miss_counter_bsz,
-                                    stride_mask_cache_miss_counter_head_kv,
-                                    stride_mask_cache_miss_counter_tsrc,
-                                    queries,
-                                    idx_bsz,
-                                    idx_tsrc,
-                                    idx_head // HEAD_GROUP,
-                                    idx_hid,
-                                    idx_chunk,
-                                    mask_tsrc_active,
-                                    mask_tdst,
-                                    mask_hid,
-                                    real_pos_tdst_min,
-                                    model_context_length,
-                                    num_sinks,
-                                    USING_EXTEND,
-                                    EXTEND_BACKEND,
-                                    NEED_APPLY_ROPE,
-                                    BLOCK_CHUNK,
-                                    BLOCK_HID,
-                                    HID_DIM,
-                                    False,
-                                    HEAD // HEAD_GROUP,
-                                    UPDATE_CACHE,
-                                    rope_range_begin,
-                                    rope_range_end,
-                                )
-
-                                t_scores_left = tl.dot(
-                                    (
-                                        queries
-                                        * (
-                                            tl.sqrt(HID_DIM * 1.0)
-                                            / tl.sqrt(tl.sqrt(HID_DIM * 1.0))
-                                        ).to(queries.dtype)
-                                    ).to(queries.dtype),
-                                    (
-                                        keys_left.to(queries.dtype)
-                                        * (1 / tl.sqrt(tl.sqrt(HID_DIM * 1.0))).to(
-                                            queries.dtype
-                                        )
-                                    ).to(queries.dtype),
-                                    allow_tf32=True,
-                                    out_dtype=tl.float32,
-                                ).to(queries.dtype)
-
-                                scores_left = tl.maximum(scores_left, t_scores_left)
-                        else:
-                            idx_tsrc = (idx_tsrc_left + idx_tsrc_center) // 2
-                            keys_left = load_keys_with_rope(
-                                K,
-                                stride_k_bsz,
-                                stride_k_tsrc,
-                                stride_k_head_kv,
-                                stride_k_hid,
-                                COS,
-                                stride_cos_t,
-                                stride_cos_hid,
-                                SIN,
-                                stride_sin_t,
-                                stride_sin_hid,
-                                # paged attention args template
-                                USING_PAGES,
-                                PAGE_SIZE,
-                                K_CACHE,
-                                stride_k_cache_page,
-                                stride_k_cache_offset,
-                                stride_k_cache_kv_head,
-                                stride_k_cache_hid,
-                                BLOCK_TABLE,
-                                stride_block_table_bsz,
-                                stride_block_table_page,
-                                CACHE_SEQ_LENS,
-                                stride_cache_seq_lens_b,
-                                USING_OFFLOAD_CACHE,
-                                OFFLOAD_CACHE_KV_PACKED,
-                                GPU_BANK_COUNT,
-                                OFFLOAD_CACHE_UVM_METADATA,
-                                stride_offload_cache_uvm_metadata_token,
-                                stride_offload_cache_uvm_metadata_k,
-                                OFFLOAD_CACHE_GPU_GLOBAL_METADATA,
-                                stride_offload_cache_gpu_global_metadata_k,
-                                stride_offload_cache_gpu_global_metadata_pad,
-                                OFFLOAD_CACHE_GPU_BANK,
-                                stride_offload_cache_gpu_bank_token,
-                                stride_offload_cache_gpu_bank_hid,
-                                OFFLOAD_CACHE_GPU_METADATA,
-                                stride_offload_cache_gpu_metadata_token,
-                                stride_offload_cache_gpu_metadata_k,
-                                OFFLOAD_CACHE_GPU_TABLE,
-                                stride_offload_cache_gpu_table_head_kv,
-                                stride_offload_cache_gpu_table_token,
-                                strdie_offload_cache_gpu_table_k,
-                                MASK_ACCESS_COUNTER,
-                                stride_mask_access_counter_bsz,
-                                stride_mask_access_counter_head_kv,
-                                stride_mask_access_counter_tsrc,
-                                MASK_CACHE_MISS_COUNTER,
-                                stride_mask_cache_miss_counter_bsz,
-                                stride_mask_cache_miss_counter_head_kv,
-                                stride_mask_cache_miss_counter_tsrc,
-                                queries,
-                                idx_bsz,
-                                idx_tsrc,
-                                idx_head // HEAD_GROUP,
-                                idx_hid,
-                                idx_chunk,
-                                mask_tsrc_active,
-                                mask_tdst,
-                                mask_hid,
-                                real_pos_tdst_min,
-                                model_context_length,
-                                num_sinks,
-                                USING_EXTEND,
-                                EXTEND_BACKEND,
-                                NEED_APPLY_ROPE,
-                                BLOCK_CHUNK,
-                                BLOCK_HID,
-                                HID_DIM,
-                                False,
-                                HEAD // HEAD_GROUP,
-                                UPDATE_CACHE,
-                                rope_range_begin,
-                                rope_range_end,
-                            )
-
-                            scores_left = tl.dot(
-                                (
-                                    queries
-                                    * (
-                                        tl.sqrt(HID_DIM * 1.0)
-                                        / tl.sqrt(tl.sqrt(HID_DIM * 1.0))
-                                    ).to(queries.dtype)
-                                ).to(queries.dtype),
-                                (
-                                    keys_left.to(queries.dtype)
-                                    * (1 / tl.sqrt(tl.sqrt(HID_DIM * 1.0))).to(
-                                        queries.dtype
-                                    )
-                                ).to(queries.dtype),
-                                allow_tf32=True,
-                                out_dtype=tl.float32,
-                            ).to(queries.dtype)
-
-                        if REDUCE == "max":
-                            scores_left = tl.where(
-                                mask_tdst[:, None], scores_left, float("-inf")
-                            )
-                            scores_left = tl.max(scores_left, axis=0).to(
-                                scores_left.dtype
-                            )
-                        elif REDUCE == "mean":
-                            scores_left = tl.where(
-                                mask_tdst[:, None], scores_left, float("0")
-                            )
-                            scores_left = tl.sum(scores_left, axis=0).to(
-                                scores_left.dtype
-                            )
-                            scores_left = (
-                                scores_left / tl.sum(mask_tdst.to(tl.float32))
-                            ).to(scores_left.dtype)
-                        else:
-                            raise Exception()
-                        scores_left = tl.where(
-                            mask_tsrc_active, scores_left, float("-inf")
-                        ).to(scores_left.dtype)
-
-                        if ORACLE_MAXIMUM:
-                            scores_right = tl.full(
-                                (idx_tdst.shape[0], BLOCK_CHUNK),
-                                dtype=tl.float32,
-                                value=float("-inf"),
-                            )
-                            for tt in range(0, max_chunk_size.to(tl.int32)):
-                                idx_tsrc = idx_tsrc_center + tt
-                                keys_right = load_keys_with_rope(
-                                    K,
-                                    stride_k_bsz,
-                                    stride_k_tsrc,
-                                    stride_k_head_kv,
-                                    stride_k_hid,
-                                    COS,
-                                    stride_cos_t,
-                                    stride_cos_hid,
-                                    SIN,
-                                    stride_sin_t,
-                                    stride_sin_hid,
-                                    # paged attention args template
-                                    USING_PAGES,
-                                    PAGE_SIZE,
-                                    K_CACHE,
-                                    stride_k_cache_page,
-                                    stride_k_cache_offset,
-                                    stride_k_cache_kv_head,
-                                    stride_k_cache_hid,
-                                    BLOCK_TABLE,
-                                    stride_block_table_bsz,
-                                    stride_block_table_page,
-                                    CACHE_SEQ_LENS,
-                                    stride_cache_seq_lens_b,
-                                    USING_OFFLOAD_CACHE,
-                                    OFFLOAD_CACHE_KV_PACKED,
-                                    GPU_BANK_COUNT,
-                                    OFFLOAD_CACHE_UVM_METADATA,
-                                    stride_offload_cache_uvm_metadata_token,
-                                    stride_offload_cache_uvm_metadata_k,
-                                    OFFLOAD_CACHE_GPU_GLOBAL_METADATA,
-                                    stride_offload_cache_gpu_global_metadata_k,
-                                    stride_offload_cache_gpu_global_metadata_pad,
-                                    OFFLOAD_CACHE_GPU_BANK,
-                                    stride_offload_cache_gpu_bank_token,
-                                    stride_offload_cache_gpu_bank_hid,
-                                    OFFLOAD_CACHE_GPU_METADATA,
-                                    stride_offload_cache_gpu_metadata_token,
-                                    stride_offload_cache_gpu_metadata_k,
-                                    OFFLOAD_CACHE_GPU_TABLE,
-                                    stride_offload_cache_gpu_table_head_kv,
-                                    stride_offload_cache_gpu_table_token,
-                                    strdie_offload_cache_gpu_table_k,
-                                    MASK_ACCESS_COUNTER,
-                                    stride_mask_access_counter_bsz,
-                                    stride_mask_access_counter_head_kv,
-                                    stride_mask_access_counter_tsrc,
-                                    MASK_CACHE_MISS_COUNTER,
-                                    stride_mask_cache_miss_counter_bsz,
-                                    stride_mask_cache_miss_counter_head_kv,
-                                    stride_mask_cache_miss_counter_tsrc,
-                                    queries,
-                                    idx_bsz,
-                                    idx_tsrc,
-                                    idx_head // HEAD_GROUP,
-                                    idx_hid,
-                                    idx_chunk,
-                                    mask_tsrc_active,
-                                    mask_tdst,
-                                    mask_hid,
-                                    real_pos_tdst_min,
-                                    model_context_length,
-                                    num_sinks,
-                                    USING_EXTEND,
-                                    EXTEND_BACKEND,
-                                    NEED_APPLY_ROPE,
-                                    BLOCK_CHUNK,
-                                    BLOCK_HID,
-                                    HID_DIM,
-                                    True,
-                                    HEAD // HEAD_GROUP,
-                                    UPDATE_CACHE,
-                                    rope_range_begin,
-                                    rope_range_end,
-                                )
-
-                                t_scores_right = tl.dot(
-                                    (
-                                        queries
-                                        * (
-                                            tl.sqrt(HID_DIM * 1.0)
-                                            / tl.sqrt(tl.sqrt(HID_DIM * 1.0))
-                                        ).to(queries.dtype)
-                                    ).to(queries.dtype),
-                                    (
-                                        keys_right.to(queries.dtype)
-                                        * (1 / tl.sqrt(tl.sqrt(HID_DIM * 1.0))).to(
-                                            queries.dtype
-                                        )
-                                    ).to(queries.dtype),
-                                    allow_tf32=True,
-                                    out_dtype=tl.float32,
-                                ).to(queries.dtype)
-
-                                scores_right = tl.maximum(scores_right, t_scores_right)
-                        else:
-                            idx_tsrc = (idx_tsrc_center + idx_tsrc_right) // 2
-                            keys_right = load_keys_with_rope(
-                                K,
-                                stride_k_bsz,
-                                stride_k_tsrc,
-                                stride_k_head_kv,
-                                stride_k_hid,
-                                COS,
-                                stride_cos_t,
-                                stride_cos_hid,
-                                SIN,
-                                stride_sin_t,
-                                stride_sin_hid,
-                                # paged attention args template
-                                USING_PAGES,
-                                PAGE_SIZE,
-                                K_CACHE,
-                                stride_k_cache_page,
-                                stride_k_cache_offset,
-                                stride_k_cache_kv_head,
-                                stride_k_cache_hid,
-                                BLOCK_TABLE,
-                                stride_block_table_bsz,
-                                stride_block_table_page,
-                                CACHE_SEQ_LENS,
-                                stride_cache_seq_lens_b,
-                                USING_OFFLOAD_CACHE,
-                                OFFLOAD_CACHE_KV_PACKED,
-                                GPU_BANK_COUNT,
-                                OFFLOAD_CACHE_UVM_METADATA,
-                                stride_offload_cache_uvm_metadata_token,
-                                stride_offload_cache_uvm_metadata_k,
-                                OFFLOAD_CACHE_GPU_GLOBAL_METADATA,
-                                stride_offload_cache_gpu_global_metadata_k,
-                                stride_offload_cache_gpu_global_metadata_pad,
-                                OFFLOAD_CACHE_GPU_BANK,
-                                stride_offload_cache_gpu_bank_token,
-                                stride_offload_cache_gpu_bank_hid,
-                                OFFLOAD_CACHE_GPU_METADATA,
-                                stride_offload_cache_gpu_metadata_token,
-                                stride_offload_cache_gpu_metadata_k,
-                                OFFLOAD_CACHE_GPU_TABLE,
-                                stride_offload_cache_gpu_table_head_kv,
-                                stride_offload_cache_gpu_table_token,
-                                strdie_offload_cache_gpu_table_k,
-                                MASK_ACCESS_COUNTER,
-                                stride_mask_access_counter_bsz,
-                                stride_mask_access_counter_head_kv,
-                                stride_mask_access_counter_tsrc,
-                                MASK_CACHE_MISS_COUNTER,
-                                stride_mask_cache_miss_counter_bsz,
-                                stride_mask_cache_miss_counter_head_kv,
-                                stride_mask_cache_miss_counter_tsrc,
-                                queries,
-                                idx_bsz,
-                                idx_tsrc,
-                                idx_head // HEAD_GROUP,
-                                idx_hid,
-                                idx_chunk,
-                                mask_tsrc_active,
-                                mask_tdst,
-                                mask_hid,
-                                real_pos_tdst_min,
-                                model_context_length,
-                                num_sinks,
-                                USING_EXTEND,
-                                EXTEND_BACKEND,
-                                NEED_APPLY_ROPE,
-                                BLOCK_CHUNK,
-                                BLOCK_HID,
-                                HID_DIM,
-                                True,
-                                HEAD // HEAD_GROUP,
-                                UPDATE_CACHE,
-                                rope_range_begin,
-                                rope_range_end,
-                            )
-
-                            scores_right = tl.dot(
-                                (
-                                    queries
-                                    * (
-                                        tl.sqrt(HID_DIM * 1.0)
-                                        / tl.sqrt(tl.sqrt(HID_DIM * 1.0))
-                                    ).to(queries.dtype)
-                                ).to(queries.dtype),
-                                (
-                                    keys_right.to(queries.dtype)
-                                    * (1 / tl.sqrt(tl.sqrt(HID_DIM * 1.0))).to(
-                                        queries.dtype
-                                    )
-                                ).to(queries.dtype),
-                                allow_tf32=True,
-                                out_dtype=tl.float32,
-                            ).to(queries.dtype)
-
-                        if REDUCE == "max":
-                            scores_right = tl.where(
-                                mask_tdst[:, None], scores_right, float("-inf")
-                            )
-                            scores_right = tl.max(scores_right, axis=0).to(
-                                scores_right.dtype
-                            )
-                        elif REDUCE == "mean":
-                            scores_right = tl.where(
-                                mask_tdst[:, None], scores_right, float("0")
-                            )
-                            scores_right = tl.sum(scores_right, axis=0).to(
-                                scores_right.dtype
-                            )
-                            scores_right = (
-                                scores_right / tl.sum(mask_tdst.to(tl.float32))
-                            ).to(scores_right.dtype)
-                        else:
-                            raise Exception()
-                        scores_right = tl.where(
-                            mask_tsrc_active, scores_right, float("-inf")
-                        ).to(scores_right.dtype)
-
-                        mask_left_win = scores_left > scores_right
-                        idx_tsrc_left = tl.where(
-                            mask_tsrc_active,
-                            tl.where(
-                                mask_left_win,
-                                idx_tsrc_left,
-                                idx_tsrc_center,
-                            ),
-                            idx_tsrc_left,
-                        )
-
-                        idx_tsrc_right = tl.where(
-                            mask_tsrc_active,
-                            tl.where(
-                                mask_left_win,
-                                idx_tsrc_center,
-                                idx_tsrc_right,
-                            ),
-                            idx_tsrc_right,
-                        )
-
-                        scores = tl.maximum(
-                            scores,
-                            tl.where(
-                                mask_tsrc_active,
-                                tl.where(
-                                    mask_left_win,
-                                    scores_left,
-                                    scores_right,
-                                ),
-                                scores,
-                            ),
-                        )
-
-                    # idx_tsrc_center = (idx_tsrc_left + idx_tsrc_right) // 2
-                    # idx_tsrc_left = idx_tsrc_center - TERMINATE_SIZE // 2
-                    # idx_tsrc_right = idx_tsrc_left + TERMINATE_SIZE
-
-                    tl.store(
-                        INDICES_LEFT
-                        + idx_bsz * stride_indices_left_bsz
-                        + idx_bdst_scan * stride_indices_left_bdst
-                        + idx_head * stride_indices_left_head
-                        + idx_chunk * stride_indices_left_chunk,
-                        value=idx_tsrc_left,
-                        mask=mask_chunk,
-                    )
-
-                    tl.store(
-                        INDICES_RIGHT
-                        + idx_bsz * stride_indices_right_bsz
-                        + idx_bdst_scan * stride_indices_right_bdst
-                        + idx_head * stride_indices_right_head
-                        + idx_chunk * stride_indices_right_chunk,
-                        value=idx_tsrc_right,
-                        mask=mask_chunk,
-                    )
-
-                    tl.store(
-                        OUT_SCORES
-                        + idx_bsz * stride_out_scores_bsz
-                        + idx_bdst_scan * stride_out_scores_bdst
-                        + idx_head * stride_out_scores_head
-                        + idx_chunk * stride_out_scores_chunk,
-                        value=scores,
-                        mask=mask_chunk,
-                    )
-
-
-@triton.jit
-def calculate_chunk_score(
-    Q,
-    stride_q_bsz,
-    stride_q_tdst,
-    stride_q_head,
-    stride_q_hid,
-    K,
-    stride_k_bsz,
-    stride_k_tsrc,
-    stride_k_head_kv,
-    stride_k_hid,
-    POS,
-    stride_pos_bsz,
-    stride_pos_tdst,
-    COS,
-    stride_cos_t,
-    stride_cos_hid,
-    SIN,
-    stride_sin_t,
-    stride_sin_hid,
-    # paged attention args template
-    USING_PAGES: tl.constexpr,
-    PAGE_SIZE: tl.constexpr,
-    K_CACHE,
-    stride_k_cache_page,
-    stride_k_cache_offset,
-    stride_k_cache_kv_head,
-    stride_k_cache_hid,
-    V_CACHE,
-    stride_v_cache_page,
-    stride_v_cache_offset,
-    stride_v_cache_kv_head,
-    stride_v_cache_hid,
-    BLOCK_TABLE,
-    stride_block_table_bsz,
-    stride_block_table_page,
-    CACHE_SEQ_LENS,
-    stride_cache_seq_lens_b,
-    # offload cache args template
-    USING_OFFLOAD_CACHE: tl.constexpr,
-    OFFLOAD_CACHE_METHOD: tl.constexpr,
-    OFFLOAD_CACHE_BUDGET: tl.constexpr,
-    OFFLOAD_CACHE_KV_HEAD: tl.constexpr,
-    OFFLOAD_CACHE_K_TABLES,
-    stride_offload_cache_k_tables_n,
-    stride_offload_cache_k_tables_t,
-    OFFLOAD_CACHE_K_BANKS,
-    stride_offload_cache_k_banks_n,
-    stride_offload_cache_k_banks_page,
-    stride_offload_cache_k_banks_offset,
-    stride_offload_cache_k_banks_hid,
-    OFFLOAD_CACHE_K_BANK_STATS,
-    stride_offload_cache_k_bank_stats_n,
-    stride_offload_cache_k_bank_stats_page,
-    stride_offload_cache_k_bank_stats_k,
-    OFFLOAD_CACHE_COUNTERS,
-    stride_offload_cache_counters_n,
-    stride_offload_cache_counters_k,
-    INDICES_LEFT,
-    stride_indices_left_bsz,
-    stride_indices_left_bdst,
-    stride_indices_left_head,
-    stride_indices_left_chunk,
-    INDICES_RIGHT,
-    stride_indices_right_bsz,
-    stride_indices_right_bdst,
-    stride_indices_right_head,
-    stride_indices_right_chunk,
-    OUT_SCORES,
-    stride_out_scores_bsz,
-    stride_out_scores_bdst,
-    stride_out_scores_head,
-    stride_out_scores_chunk,
-    model_context_length,
-    sliding_window_size,
-    num_sinks,
-    max_chunk_size,
-    TDST,
-    BDST,
-    BDST_SCAN,
-    N_HEAD,
-    N_CHUNK,
-    HEAD_GROUP,
-    USING_EXTEND: tl.constexpr,
-    NEED_APPLY_ROPE: tl.constexpr,
-    EXTEND_BACKEND: tl.constexpr,
-    BLOCK_HID: tl.constexpr,
-    BLOCK_SIZE_Q: tl.constexpr,
-    BLOCK_STRIDE_Q: tl.constexpr,
-    BLOCK_SIZE_K: tl.constexpr,
-    BLOCK_STRIDE_K: tl.constexpr,
-    SCAN_STRIDE: tl.constexpr,
-    BLOCK_CHUNK: tl.constexpr,
-    REDUCE: tl.constexpr = "max",
-):
-    pid = tl.program_id(0).to(tl.int64)
-
-    # idx_chunk = pid % N_CHUNK
-    # pid = pid // N_CHUNK
-    idx_head = pid % N_HEAD
-    pid = pid // N_HEAD
-    idx_bdst_scan = pid % BDST_SCAN
-    pid = pid // BDST_SCAN
-    idx_bsz = pid
-
-    tl.static_assert(
-        (NEED_APPLY_ROPE and USING_EXTEND) or (not (NEED_APPLY_ROPE or USING_EXTEND))
-    )
-
-    idx_tdst = (
-        idx_bdst_scan * SCAN_STRIDE * BLOCK_SIZE_Q
-        + (BDST * BLOCK_SIZE_Q - BDST_SCAN * SCAN_STRIDE * BLOCK_SIZE_Q)
-        + tl.arange(0, BLOCK_SIZE_Q // BLOCK_STRIDE_Q) * BLOCK_STRIDE_Q
-    )
-    mask_tdst = (idx_tdst < TDST) & (idx_tdst >= 0)
-    idx_hid = tl.arange(0, BLOCK_HID)
-    mask_hid = idx_hid < BLOCK_HID
-
-    pos_tdst = tl.load(
-        POS + idx_bsz * stride_pos_bsz + idx_tdst * stride_pos_tdst,
-        mask=mask_tdst,
-        other=0,
-    )
-    pos_tdst_min = tl.min(tl.where(mask_tdst, pos_tdst, 999999999))
-    pos_tdst_max = tl.max(pos_tdst)
-
-    # real_pos_tdst_min = idx_bdst * BLOCK_SIZE_Q + TSRC - TDST
-    # real_pos_tdst_min = tl.min(tl.where(mask_tdst, pos_tdst, 99999999999))
-
-    # pos_tdst_min = (real_pos_tdst_min - sliding_window_size - num_sinks).to(tl.int32)
-    # pos_tdst_min = tl.maximum(pos_tdst_min, 0)
-
-    queries = tl.load(
-        Q
-        + idx_bsz * stride_q_bsz
-        + idx_tdst[:, None] * stride_q_tdst
-        + idx_head * stride_q_head
-        + idx_hid[None, :] * stride_q_hid,
-        mask=mask_tdst[:, None],
-        other=0,
-    )
-
-    if NEED_APPLY_ROPE and USING_EXTEND:
-        if EXTEND_BACKEND == "dynamic_extend":
-            new_tdst = pos_tdst
-        elif EXTEND_BACKEND == "self_extend":
-            new_tdst = pos_tdst
-        elif EXTEND_BACKEND == "streaming":
-            new_tdst = tl.minimum(pos_tdst, N_CHUNK + sliding_window_size)
-        elif EXTEND_BACKEND == "relative":
-            new_tdst = pos_tdst * 0 + sliding_window_size
-        else:
-            raise Exception()
-
-        queries_rot = tl.load(
-            Q
-            + idx_bsz * stride_q_bsz
-            + idx_tdst[:, None] * stride_q_tdst
-            + idx_head * stride_q_head
-            + ((idx_hid + BLOCK_HID // 2) % BLOCK_HID)[None, :] * stride_q_hid,
-            mask=mask_tdst[:, None],
-            other=0,
-        )
-
-        cos_new = tl.load(
-            COS
-            + new_tdst[:, None].to(tl.int64) * stride_cos_t
-            + (idx_hid % (BLOCK_HID // 2))[None, :] * stride_cos_hid,
-            mask=mask_tdst[:, None],
-            other=0.0,
-        ).to(queries.dtype)
-        sin_new = tl.load(
-            SIN
-            + new_tdst[:, None].to(tl.int64) * stride_sin_t
-            + (idx_hid % (BLOCK_HID // 2))[None, :] * stride_sin_hid,
-            mask=mask_tdst[:, None],
-            other=0.0,
-        ).to(queries.dtype)
-
-        queries_rot = queries_rot * (
-            ((idx_hid + BLOCK_HID // 2)[None, :] < BLOCK_HID) * (-2) + 1
-        ).to(queries_rot.dtype)
-
-        queries = (queries * cos_new + queries_rot * sin_new).to(queries.dtype)
-
-    for idx_chunk_start in range(0, N_CHUNK, BLOCK_CHUNK):
-        # for idx_chunk in range(tl.cdiv(pos_tdst_max, max_chunk_size)):
-        idx_chunk = tl.arange(0, BLOCK_CHUNK) + idx_chunk_start
-        mask_chunk = idx_chunk < N_CHUNK
-        idx_tsrc_left = tl.load(
-            INDICES_LEFT
-            + idx_bsz * stride_indices_left_bsz
-            + idx_bdst_scan * stride_indices_left_bdst
-            + idx_head * stride_indices_left_head
-            + idx_chunk * stride_indices_left_chunk,
-            mask=mask_chunk,
-            other=987654321,
-        ).to(tl.int64)
-
-        idx_tsrc_right = tl.load(
-            INDICES_RIGHT
-            + idx_bsz * stride_indices_right_bsz
-            + idx_bdst_scan * stride_indices_right_bdst
-            + idx_head * stride_indices_right_head
-            + idx_chunk * stride_indices_right_chunk,
-            mask=mask_chunk,
-            other=987654321,
-        ).to(tl.int64)
-
-        if tl.min(idx_tsrc_left) <= pos_tdst_max:
-            idx_tsrc_center = (idx_tsrc_left + idx_tsrc_right) // 2
-            idx_tsrc_left = tl.maximum(0, idx_tsrc_center - BLOCK_SIZE_K // 2)
-            idx_tsrc = (
-                idx_tsrc_left[:, None]
-                + tl.arange(0, BLOCK_SIZE_K // BLOCK_STRIDE_K)[None, :] * BLOCK_STRIDE_K
-            )
-            idx_tsrc = tl.ravel(idx_tsrc)
-            mask_tsrc = idx_tsrc <= (tl.max(pos_tdst) - sliding_window_size)
-
-            keys = load_keys_with_rope(
-                K,
-                stride_k_bsz,
-                stride_k_tsrc,
-                stride_k_head_kv,
-                stride_k_hid,
-                COS,
-                stride_cos_t,
-                stride_cos_hid,
-                SIN,
-                stride_sin_t,
-                stride_sin_hid,
-                # paged attention args template
-                USING_PAGES,
-                PAGE_SIZE,
-                K_CACHE,
-                stride_k_cache_page,
-                stride_k_cache_offset,
-                stride_k_cache_kv_head,
-                stride_k_cache_hid,
-                BLOCK_TABLE,
-                stride_block_table_bsz,
-                stride_block_table_page,
-                CACHE_SEQ_LENS,
-                stride_cache_seq_lens_b,
-                # offload cache args template
-                USING_OFFLOAD_CACHE,
-                OFFLOAD_CACHE_METHOD,
-                OFFLOAD_CACHE_BUDGET,
-                OFFLOAD_CACHE_KV_HEAD,
-                OFFLOAD_CACHE_K_TABLES,
-                stride_offload_cache_k_tables_n,
-                stride_offload_cache_k_tables_t,
-                OFFLOAD_CACHE_K_BANKS,
-                stride_offload_cache_k_banks_n,
-                stride_offload_cache_k_banks_page,
-                stride_offload_cache_k_banks_offset,
-                stride_offload_cache_k_banks_hid,
-                OFFLOAD_CACHE_K_BANK_STATS,
-                stride_offload_cache_k_bank_stats_n,
-                stride_offload_cache_k_bank_stats_page,
-                stride_offload_cache_k_bank_stats_k,
-                OFFLOAD_CACHE_COUNTERS,
-                stride_offload_cache_counters_n,
-                stride_offload_cache_counters_k,
-                queries,
-                idx_bsz,
-                idx_tsrc,
-                idx_head // HEAD_GROUP,
-                idx_hid,
-                idx_chunk,
-                mask_tsrc,
-                mask_tdst,
-                mask_hid,
-                pos_tdst_min,
-                model_context_length,
-                num_sinks,
-                USING_EXTEND,
-                EXTEND_BACKEND,
-                NEED_APPLY_ROPE,
-                BLOCK_SIZE_K,
-                BLOCK_HID,
-                True,
-                HEAD // HEAD_GROUP,
-                UPDATE_CACHE,
-            )
-
-            scores = tl.dot(
-                (
-                    queries
-                    * (tl.sqrt(BLOCK_HID * 1.0) / tl.sqrt(tl.sqrt(BLOCK_HID * 1.0))).to(
-                        queries.dtype
-                    )
-                ).to(queries.dtype),
-                (
-                    keys.to(queries.dtype)
-                    * (1 / tl.sqrt(tl.sqrt(BLOCK_HID * 1.0))).to(queries.dtype)
-                ).to(queries.dtype),
-                allow_tf32=True,
-                out_dtype=tl.float32,
-            ).to(queries.dtype)
-
-            if REDUCE == "max":
-                scores_reduced = tl.where(
-                    mask_tdst[:, None] & mask_tsrc[None, :], scores, -32000.0
-                )
-                scores_reduced = tl.reshape(
-                    scores_reduced,
-                    BLOCK_SIZE_Q // BLOCK_STRIDE_Q,
-                    BLOCK_CHUNK,
-                    BLOCK_SIZE_K // BLOCK_STRIDE_K,
-                )
-                scores_reduced = tl.max(scores_reduced, axis=0)
-                scores_reduced = tl.max(scores_reduced, axis=-1)
-            # elif REDUCE == 'mean':
-            #     scores_reduced = tl.sum(tl.where(
-            #         mask_tdst[:, None] & mask_tsrc[None, :],
-            #         scores,
-            #         0
-            #     )) / tl.sum((mask_tdst[:, None] & mask_tsrc[None, :]).to(tl.int32))
-            else:
-                raise Exception()
-
-            tl.store(
-                OUT_SCORES
-                + idx_bsz * stride_out_scores_bsz
-                + idx_bdst_scan * stride_out_scores_bdst
-                + idx_head * stride_out_scores_head
-                + idx_chunk * stride_out_scores_chunk,
-                value=scores_reduced,
-                mask=mask_chunk,
-            )
-        else:
-            tl.store(
-                OUT_SCORES
-                + idx_bsz * stride_out_scores_bsz
-                + idx_bdst_scan * stride_out_scores_bdst
-                + idx_head * stride_out_scores_head
-                + idx_chunk * stride_out_scores_chunk,
-                value=-32000.0,
-                mask=mask_chunk,
-            )
-
-
-@triton.jit
-def compute_v_cos(
-    V,
-    stride_v_bsz,
-    stride_v_tsrc,
-    stride_v_head_kv,
-    stride_v_hid,
-    INDICES,
-    stride_indices_bsz,
-    stride_indices_bdst,
-    stride_indices_head,
-    strdie_indices_k,
-    POS,
-    stride_pos_bsz,
-    stride_pos_tdst,
-    OUT_SCORES,
-    stride_out_scores_bsz,
-    stride_out_scores_bdst,
-    stride_out_scores_head,
-    stride_out_scores_k,
-    # paged attention args template
-    USING_PAGES: tl.constexpr,
-    PAGE_SIZE: tl.constexpr,
-    K_CACHE,
-    stride_k_cache_page,
-    stride_k_cache_offset,
-    stride_k_cache_kv_head,
-    stride_k_cache_hid,
-    V_CACHE,
-    stride_v_cache_page,
-    stride_v_cache_offset,
-    stride_v_cache_kv_head,
-    stride_v_cache_hid,
-    BLOCK_TABLE,
-    stride_block_table_bsz,
-    stride_block_table_page,
-    CACHE_SEQ_LENS,
-    stride_cache_seq_lens_b,
-    # offload cache args template
-    USING_OFFLOAD_CACHE: tl.constexpr,
-    OFFLOAD_CACHE_KV_PACKED: tl.constexpr,
-    GPU_BANK_COUNT,
-    OFFLOAD_CACHE_UVM_METADATA,
-    stride_offload_cache_uvm_metadata_token,
-    stride_offload_cache_uvm_metadata_k,
-    OFFLOAD_CACHE_GPU_GLOBAL_METADATA,
-    stride_offload_cache_gpu_global_metadata_k,
-    stride_offload_cache_gpu_global_metadata_pad,
-    OFFLOAD_CACHE_GPU_BANK,
-    stride_offload_cache_gpu_bank_token,
-    stride_offload_cache_gpu_bank_hid,
-    OFFLOAD_CACHE_GPU_METADATA,
-    stride_offload_cache_gpu_metadata_token,
-    stride_offload_cache_gpu_metadata_k,
-    OFFLOAD_CACHE_GPU_TABLE,
-    stride_offload_cache_gpu_table_head_kv,
-    stride_offload_cache_gpu_table_token,
-    strdie_offload_cache_gpu_table_k,
-    ACCESS_COUNTER,
-    stride_access_counter_bsz,
-    stride_access_counter_head_kv,
-    stride_access_counter_tsrc,
-    CACHE_MISS_COUNTER,
-    stride_cache_miss_counter_bsz,
-    stride_cache_miss_counter_head_kv,
-    stride_cache_miss_counter_tsrc,
-    TDST,
-    TSRC,
-    HEAD,
-    KS,
-    HEAD_GROUP: tl.constexpr,
-    GROUP_K: tl.constexpr,
-    BLOCK_SIZE_Q: tl.constexpr,
-    BLOCK_SIZE_K: tl.constexpr,
-    BLOCK_STRIDE_Q: tl.constexpr,
-    BLOCK_STRIDE_K: tl.constexpr,
-    BLOCK_HID: tl.constexpr,
-):
-    pid = tl.program_id(0)
-
-    idx_head = pid % HEAD
-    pid = pid // HEAD
-
-    idx_bk = pid % tl.cdiv(KS, GROUP_K)
-    idx_k = idx_bk * GROUP_K + tl.arange(0, GROUP_K)
-    mask_k = idx_k < KS
-    pid = pid // tl.cdiv(KS, GROUP_K)
-
-    idx_bdst = pid % tl.cdiv(TDST, BLOCK_SIZE_Q)
-    idx_tdst = (
-        idx_bdst * BLOCK_SIZE_Q
-        + tl.arange(0, BLOCK_SIZE_Q // BLOCK_STRIDE_Q) * BLOCK_STRIDE_Q
-    )
-    mask_tdst = idx_tdst < TDST
-    idx_bsz = pid // tl.cdiv(TDST, BLOCK_SIZE_Q)
-
-    idx_hid = tl.arange(0, BLOCK_HID)
-
-    pos_tdst = tl.load(
-        POS + idx_bsz * stride_pos_bsz + idx_tdst * stride_pos_tdst,
-        mask=mask_tdst,
-        other=0,
-    )
-    mask_tdst = mask_tdst & (pos_tdst < TSRC)
-    seq_len = tl.max(pos_tdst)  # + 1
-
-    indices = tl.load(
-        INDICES
-        + idx_bsz * stride_indices_bsz
-        + idx_bdst * stride_indices_bdst
-        + idx_head * stride_indices_head
-        + idx_k * strdie_indices_k,
-        mask=mask_k,
-        other=seq_len + 2 * BLOCK_SIZE_K,
-    )
-    indices = indices // BLOCK_SIZE_K * BLOCK_SIZE_K
-
-    idx_tsrc = tl.ravel(indices[:, None] + tl.arange(0, BLOCK_SIZE_K)[None, :])
-    mask_tsrc = (idx_tsrc < seq_len) & (idx_tsrc >= 0)
-
-    # values_tdst = tl.load(
-    #     V +\
-    #         idx_bsz * stride_v_bsz+\
-    #         idx_tdst[:, None] * stride_v_tsrc+\
-    #         (idx_head // HEAD_GROUP) * stride_v_head_kv +\
-    #         idx_hid[None, :] * strdie_v_hid,
-    #     mask=mask_tdst[:, None],
-    #     other=0,
-    # )
-
-    # tl.static_assert(not USING_OFFLOAD_CACHE)
-    values_tdst = load_tokens(
-        V,
-        stride_v_bsz,
-        stride_v_tsrc,
-        stride_v_head_kv,
-        stride_v_hid,
-        USING_PAGES,
-        PAGE_SIZE,
-        V_CACHE,
-        stride_v_cache_page,
-        stride_v_cache_offset,
-        stride_v_cache_kv_head,
-        stride_v_cache_hid,
-        BLOCK_TABLE,
-        stride_block_table_bsz,
-        stride_block_table_page,
-        CACHE_SEQ_LENS,
-        stride_cache_seq_lens_b,
-        USING_OFFLOAD_CACHE,
-        OFFLOAD_CACHE_KV_PACKED,
-        GPU_BANK_COUNT,
-        True,
-        OFFLOAD_CACHE_UVM_METADATA,
-        stride_offload_cache_uvm_metadata_token,
-        stride_offload_cache_uvm_metadata_k,
-        OFFLOAD_CACHE_GPU_GLOBAL_METADATA,
-        stride_offload_cache_gpu_global_metadata_k,
-        stride_offload_cache_gpu_global_metadata_pad,
-        OFFLOAD_CACHE_GPU_BANK,
-        stride_offload_cache_gpu_bank_token,
-        stride_offload_cache_gpu_bank_hid,
-        OFFLOAD_CACHE_GPU_METADATA,
-        stride_offload_cache_gpu_metadata_token,
-        stride_offload_cache_gpu_metadata_k,
-        OFFLOAD_CACHE_GPU_TABLE,
-        stride_offload_cache_gpu_table_head_kv,
-        stride_offload_cache_gpu_table_token,
-        strdie_offload_cache_gpu_table_k,
-        ACCESS_COUNTER,
-        stride_access_counter_bsz,
-        stride_access_counter_head_kv,
-        stride_access_counter_tsrc,
-        CACHE_MISS_COUNTER,
-        stride_cache_miss_counter_bsz,
-        stride_cache_miss_counter_head_kv,
-        stride_cache_miss_counter_tsrc,
-        idx_bsz,
-        pos_tdst[:, None],
-        idx_head // HEAD_GROUP,
-        idx_hid[None, :],
-        mask_tdst[:, None],
-        HEAD // HEAD_GROUP,
-        BLOCK_SIZE_Q // BLOCK_STRIDE_Q,
-        BLOCK_HID,
-    ).to(tl.bfloat16)
-
-    # values_tdst = (
-    #     tl.sum(values_tdst, axis=0) /\
-    #     tl.sum(mask_tdst.to(tl.int32))
-    # )
-
-    # values_tsrc = tl.load(
-    #     V +\
-    #         idx_bsz * stride_v_bsz +\
-    #         idx_tsrc[:, None] * stride_v_tsrc +\
-    #         (idx_head // HEAD_GROUP) * stride_v_head_kv +\
-    #         idx_hid[None, :] * strdie_v_hid,
-    #     mask=mask_tsrc[:, None],
-    #     other=0,
-    # )
-
-    values_tsrc = load_tokens(
-        V,
-        stride_v_bsz,
-        stride_v_tsrc,
-        stride_v_head_kv,
-        stride_v_hid,
-        USING_PAGES,
-        PAGE_SIZE,
-        V_CACHE,
-        stride_v_cache_page,
-        stride_v_cache_offset,
-        stride_v_cache_kv_head,
-        stride_v_cache_hid,
-        BLOCK_TABLE,
-        stride_block_table_bsz,
-        stride_block_table_page,
-        CACHE_SEQ_LENS,
-        stride_cache_seq_lens_b,
-        USING_OFFLOAD_CACHE,
-        OFFLOAD_CACHE_KV_PACKED,
-        GPU_BANK_COUNT,
-        True,
-        OFFLOAD_CACHE_UVM_METADATA,
-        stride_offload_cache_uvm_metadata_token,
-        stride_offload_cache_uvm_metadata_k,
-        OFFLOAD_CACHE_GPU_GLOBAL_METADATA,
-        stride_offload_cache_gpu_global_metadata_k,
-        stride_offload_cache_gpu_global_metadata_pad,
-        OFFLOAD_CACHE_GPU_BANK,
-        stride_offload_cache_gpu_bank_token,
-        stride_offload_cache_gpu_bank_hid,
-        OFFLOAD_CACHE_GPU_METADATA,
-        stride_offload_cache_gpu_metadata_token,
-        stride_offload_cache_gpu_metadata_k,
-        OFFLOAD_CACHE_GPU_TABLE,
-        stride_offload_cache_gpu_table_head_kv,
-        stride_offload_cache_gpu_table_token,
-        strdie_offload_cache_gpu_table_k,
-        ACCESS_COUNTER,
-        stride_access_counter_bsz,
-        stride_access_counter_head_kv,
-        stride_access_counter_tsrc,
-        CACHE_MISS_COUNTER,
-        stride_cache_miss_counter_bsz,
-        stride_cache_miss_counter_head_kv,
-        stride_cache_miss_counter_tsrc,
-        idx_bsz,
-        idx_tsrc[:, None],
-        idx_head // HEAD_GROUP,
-        idx_hid[None, :],
-        mask_tsrc[:, None],
-        HEAD // HEAD_GROUP,
-        GROUP_K * BLOCK_SIZE_K,
-        BLOCK_HID,
-    ).to(tl.bfloat16)
-
-    # values_tsrc = (
-    #     tl.sum(tl.reshape(values_tsrc, [GROUP_K, BLOCK_SIZE_K, BLOCK_HID]), axis=1) /\
-    #     tl.sum(tl.reshape(mask_tsrc.to(tl.int32), [GROUP_K, BLOCK_SIZE_K, 1]), axis=1)
-    # )
-
-    values_tdst_norm = tl.sqrt(
-        tl.sum(values_tdst.to(tl.float32) * values_tdst.to(tl.float32), axis=-1)
-    )
-    values_tsrc_norm = tl.sqrt(
-        tl.sum(values_tsrc.to(tl.float32) * values_tsrc.to(tl.float32), axis=-1)
-    )
-
-    normalized_values_tdst = values_tdst
-    normalized_values_tsrc = values_tsrc
-    normalized_values_tdst = values_tdst / tl.maximum(values_tdst_norm[:, None], 1e-20)
-    normalized_values_tsrc = values_tsrc / tl.maximum(values_tsrc_norm[:, None], 1e-20)
-
-    # -
-    # cos_sim_scores = tl.sum(normalized_values_tdst[None, :] * normalized_values_tsrc, axis=-1)
-    cos_sim_scores = tl.dot(
-        normalized_values_tdst, tl.trans(normalized_values_tsrc, 1, 0)
-    )
-    # cos_sim_scores = ((cos_sim_scores + 1) * 0.5).to(tl.float32)
-    cos_sim_scores = cos_sim_scores  # * cos_sim_scores * cos_sim_scores
-
-    scores = tl.reshape(
-        cos_sim_scores, (BLOCK_SIZE_Q // BLOCK_STRIDE_Q, GROUP_K, BLOCK_SIZE_K)
-    )
-    # scores = tl.reshape(values_tsrc_norm, (GROUP_K, BLOCK_SIZE_K))
-    mask_scores = tl.reshape(
-        mask_tdst[:, None] & mask_tsrc[None, :],
-        (BLOCK_SIZE_Q // BLOCK_STRIDE_Q, GROUP_K, BLOCK_SIZE_K),
-    )
-    scores = scores * mask_scores
-
-    # reduce-mean
-    mask_scores = tl.sum(mask_scores.to(scores.dtype), axis=-1)
-    scores = tl.sum(scores, axis=-1)
-    mask_scores = tl.sum(mask_scores.to(scores.dtype), axis=0)
-    scores = tl.sum(scores, axis=0) / tl.maximum(mask_scores, 1e-20)
-    # -
-
-    # scores = tl.sum(values_tdst[None, :] * values_tsrc, axis=1)
-
-    # reduce max
-    # scores = tl.max(tl.max(scores, axis=-1), axis=0)
-
-    # norm reduce-mean
-    # scores = tl.reshape(values_tsrc_norm, (GROUP_K, BLOCK_SIZE_K))
-    # scores = tl.sum(scores, axis=-1) / tl.maximum(tl.sum(tl.reshape(mask_tsrc, (GROUP_K, BLOCK_SIZE_K)), axis=-1), 1e-20)
-
-    # scores = tl.sum(values_tdst[None, :] * values_tsrc)
-
-    tl.store(
-        OUT_SCORES
-        + idx_bsz * stride_out_scores_bsz
-        + idx_bdst * stride_out_scores_bdst
-        + idx_head * stride_out_scores_head
-        + idx_k * stride_out_scores_k,
-        value=scores,
-        mask=mask_k,
-    )
 
 
 def dual_stage_quadratic_hip_attention(
@@ -2614,7 +759,7 @@ def dual_stage_quadratic_hip_attention(
                         group_jobs,
                         njobs,
                         HID_DIM=HID_DIM,
-                        BLOCK_HID=HID_BLOCK,  # HID_BLOCK_0=HID_BLOCK,
+                        HID_BLOCK_0=HID_BLOCK,
                         BLOCK_SIZE_Q=BLOCK_SIZE_Q,
                         STRIDE_Q=stage_block_stride_q,
                         BLOCK_CHUNK=BLOCK_CHUNK,
